huawei traces example update
